seperate Home component from App
